[Delivers #25163355] added variables now the Firefly light casting mechanics can be played with easily. In ground.js there is: bugBrightness, bugDecayPower,playerDecayPower, and willDecayPower
[#23204221] Edited will-o-wisp to "shine" darkness.
the Will-o-wisp now does not draw anything but is solely defined
by the lack of light around it. This is not meant to be the
actual implementation, since yes, we discussed having the
will-o-wisp be something and the "darkness" be something else.
I just wanted to experiment with the darkness.
The will-o-wisps are disabled by default and you need to go
into firefly.js and edit the boolean "useWills" to enable them.
There are gameplay isuses with how it is implemented now
but those can be discussed later esp since this is for
experimentation only.